Feat(#135) : 커뮤니티 상세 페이지 - 플레이어에 노래 추가 이동 Create => Contents
