+++ src/components/communityDetail/CommunityCreate.tsx
@@ -4,35 +4,21 @@ import React, { FormEvent, MouseEvent } from 'react'
 import { useSession } from 'next-auth/react'
 import { useRouter } from 'next/navigation'
 import MusicSearch from '../search/MusicSearch'
-import {
-  useCoummunityCreateItem,
-  useCoummunityItem,
-} from '@/query/communityDetail/communityMutation'
+import { useCoummunityItem } from '@/query/communityDetail/communityMutation'
 import useInput from '@/hooks/useInput'
 import Image from 'next/image'
-import { useQuery } from '@tanstack/react-query'
-import { getCurrentMusicData } from '@/shared/main/api'
 import { useMusicSearchedStore } from '@/shared/store/communityDetailStore'
 import defaultImg from '@/../public/images/plz-music.png'
 
 const CommunityCreate = () => {
   const router = useRouter()
   const { data: userSessionInfo } = useSession()
-  const uid = userSessionInfo?.user?.uid as string
   const { addCommunityMutation } = useCoummunityItem()
   const { chooseMusic } = useMusicSearchedStore()
   const musicId = chooseMusic?.musicId as string
   const musicTitle = chooseMusic?.musicTitle
   const artist = chooseMusic?.artist
   const thumbnail = chooseMusic?.thumbnail
-
-  const { updateMutation, insertMutation } = useCoummunityCreateItem()
-
-  const { data: playListCurrent } = useQuery({
-    queryFn: () => getCurrentMusicData(uid),
-    queryKey: ['playListCurrent'],
-    enabled: !!uid,
-  })
 
   const {
     form: communityForm,
@@ -66,34 +52,6 @@ const CommunityCreate = () => {
     }
   }
 
-  const onAddPlayerHandler = (
-    e: MouseEvent,
-    userId: string,
-    musicId: string,
-  ) => {
-    e.preventDefault()
-    if (!userId) {
-      alert(
-        '로그인 후 사용할 수 있는 서비스입니다. 로그인 페이지로 이동합니다.',
-      )
-      router.replace('/login')
-      return
-    }
-
-    if (playListCurrent && playListCurrent.length > 0) {
-      const currentList = playListCurrent[0].currentMusicIds
-      if (currentList.find((el) => el === musicId)) {
-        alert('이미 추가된 노래입니다.')
-        return
-      }
-      currentList.push(musicId)
-      updateMutation.mutate({ userId: uid, currentList })
-    } else {
-      insertMutation.mutate({ userId: uid, musicId })
-    }
-    alert('현재 재생목록에 추가 되었습니다.')
-  }
-
   return (
     <div>
       <form onSubmit={onSumitHandler}>
@@ -119,7 +77,7 @@ const CommunityCreate = () => {
               name='boardTitle'
               value={boardTitle}
               onChange={onChangeHandler}
-              className='mb-4 border w-full p-2 rounded-lg focus:outline-todayPink'
+              className='focus:outline-todayPink mb-4 w-full rounded-lg border p-2'
               placeholder='제목을 입력하세요'
             />
           </div>
@@ -128,7 +86,7 @@ const CommunityCreate = () => {
             name='content'
             value={content}
             onChange={onChangeHandler}
-            className='mb-4 border w-full p-2 rounded-lg focus:outline-todayPink'
+            className='focus:outline-todayPink mb-4 w-full rounded-lg border p-2'
             placeholder='내용을 입력하세요'
           />
         </div>
@@ -158,11 +116,6 @@ const CommunityCreate = () => {
           <div>
             <p>{musicTitle}</p>
             <p>{artist}</p>
-          </div>
-          <div>
-            <button onClick={(e) => onAddPlayerHandler(e, uid, musicId)}>
-              플레이어에 음악추가
-            </button>
           </div>
         </section>
       </article>
